Enabled the automaton to recv packets from a filedescriptor
I am using the Automaton to read and write packets from a named pipe rather than a network socket.
This change was necessary in order to make it work rather flawlessly, because file objects do not
have a "recv" method.
This change shouldn't change the previous behaviour because "recv" will still be called unless the
fd is an int or a "read" attribute exists.