Improved Net address generator
Net now rejects numbers above 255 and returns a len value.

Iteration is fixed to increment the octets in lowest to highest order.  Setting "iterhigh" to True will restore the old behavior of highest to lowest octets.

Adding "v" to the start of the net string will generate only "valid" addresses, by excluding 0, 127, and 224-256 from the first octet.

Internally, the "parsed" value now contains full lists rather than range tuples.  This is necessary for the "valid" functionality and is generally easier to work with.