Improved FlagsField names input
FlagsField can now use a names list longer (will be truncated) or shorter (will be filled) than the field size.  Any empty strings or otherwise missing values are assumed to be reserved, and are given the name "resX" where X is the position of the flag.  This gives an easy and consistent way to specify fields containing reserved flags, with empty strings as placeholders that can be filled once they are defined in the protocol.